Split color option functions from Cosmetics.py to Colors.py
Needed to resolve upcoming circular dependence with cosmetic plando, also cleans up Cosmetics.py a little, allowing it to focus on just the actual patching.
